% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_map.R
\name{leaf_base_lev3}
\alias{leaf_base_lev3}
\title{Base leaflet map for level 3}
\usage{
leaf_base_lev3(max_zoom = 15)
}
\arguments{
\item{max_zoom}{numeric max zoom allowed}
}
\value{
leaflet object
}
\description{
make the base map throughout the whole app
}
\examples{
\dontrun{
leaf_base_lev3 ()
}
base leaflet map for level 3
}
\author{
Maddalena Tigli
}
