% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_map.R
\name{legend_names}
\alias{legend_names}
\title{name legend}
\usage{
legend_names(layer)
}
\arguments{
\item{layer}{name of layer displayed}
}
\value{
string with title for the legend
}
\description{
conversion from name displayed to "human" name
}
\examples{
\dontrun{
legend_names ("SSP1_LandCover_1970")
}
}
\author{
Maddalena Tigli
}
